test(selecao-e2e): separa os testes que dependem de uma largura específica
Dois testes precisavam de uma viewport que a matriz do DS não oferece — o teto
de largura só é observável acima de 1200 px, e a transição da lista de etapas
exige começar estreito e alargar. Dentro da matriz, que fixa a viewport por
project, isso obrigava a ignorar o teste nos demais projects.

Passam a viver em arquivo próprio, no project autenticado, controlando o
próprio redimensionamento. Rodam uma vez em vez de doze, sem anotação de
ignorado e sem condicional.
